Fix set property error of undefined
This error occured, when Sumo API returned error or warning in
results list, eg. "Too many time series matched ...".
Variable `target` was `undefined`, because `targets.length < response.length`.
Equal lenght of both arrays is not guaranteed by the Sumo API.

The error message is saved in `errors` array and will be thrown later,
so there is no need to provide additional error handling or saving the error
with target.